requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.5.0
PyHive==0.7.0
# sasl==0.3.1
pure-sasl
thrift==0.21.0
thrift-sasl==0.4.3
requests==2.32.3
cryptography==43.0.1
GitPython==3.1.43